TST: move deprecated tests of loffset and base into test_deprecated
